+++ rules/collect.smk
@@ -11,7 +11,6 @@ localrules:
     prepare_sector_networks,
     solve_elec_networks,
     solve_sector_networks,
-    plot_networks,
 
 
 rule cluster_networks:
@@ -65,15 +64,6 @@ rule solve_sector_networks_perfect:
         expand(
             RESULTS
             + "postnetworks/elec_s{simpl}_{clusters}_l{ll}_{opts}_{sector_opts}_brownfield_all_years.nc",
-            **config["scenario"]
-        ),
-
-
-rule plot_networks:
-    input:
-        expand(
-            RESULTS
-            + "maps/elec_s{simpl}_{clusters}_l{ll}_{opts}_{sector_opts}-costs-all_{planning_horizons}.pdf",
             **config["scenario"]
         ),
